+++ liquid_extra/filters/__init__.py
@@ -3,3 +3,7 @@ from .additional import Translate
 from .array import Index
 from .html import StylesheetTag
 from .html import ScriptTag
+
+from .array import index
+from .html import stylesheet_tag
+from .html import script_tag